TempStorage: Generic refactoring of TokenStorage
The Tempstorage module holds an ``initialized`` promise that is resolved
once the data has been retrieved once from the worker. When that's done,
all data is directly accessible through getters/setters.

Data that is set in one tab/window through the modules getters/setters will
be accessible in other tabs/windows as well.

Change-Id: I7011995954fe362f6e45d2999fb998e5029cc769