Update patch set 1
Patch Set 1: Code-Review-1

(4 comments)

The ironic API is called on every pane switch in the nodes overview. I know this is how it was done in tuskar UI, but I wonder if it's enough to load it only once every time the node overview is accessed and let the tabs just filter the existing result.

Patch-set: 1