Update patch set 1
Patch Set 1:

(2 comments)

As I mention in the inline comment, we'll need to figure out this later when tripleo-common supports nodes listing. I originally was doing the call only on Nodes component mounting, but then I found it would be nice to get up to date data on tab switch. Again, It is probably just temporary as tuskar-api v3 would eventually use websockets to provide data, so we'wouldn't have to deal with this.

Patch-set: 1