Environment configuration
Environment configuration is a first step in deployment design
process, it takes information about environments provided by mapping
file retrieved from plan files to render a environment selection topics.
The output of the form which is sent to API is list of environment
filenames with information if they are used or not.

Environments are grouped by topic and each topic includes environment
groups. Environments in Environment group are mutually exclusive, so
checking one automatically disables others in the group.

Topics are displayed as panels, panels are arranged in css columns
rather than bootstrap grid to achieve Masonry-like layout.

The routing and tabbed layout is temporary and most probably the
Environment Configuration will be moved to NavBar as well as Roles
and Parameters. We need to get Plans management sorted for this
first as this configuration is happening on selected Plan.

Change-Id: I1cf99acd35723c0df08cc506d71aed193e132717