Plan Parameters listing and updating
Currently we are showing just Root template parameters, code needs
to be updated to show necessary nested ones. The list of parameters is
relatively large. This is caused by the structure of root template in
THT. It would be nice to split optional parameters/services into
separate subtemplates and activate them by environments.

The requests for listing and updating parameters take long time. We
should try to optimize somehow, but It seems the delay is rather on
tripleo-common side.

Change-Id: Icdc682fbec4a713fda78e75f6f940ac4977be4e4