Update patch set 5
Patch Set 5:

As discussed on IRC, the plan detection now lives in the PlansStore module.

Instead of calling PlansActions.choosePlan after detection, the new plan is silently set inside the store. 

If there was a previous plan that doesn't exist anymore, instead of calling NotificationActions.notify to notify the user (which would be within a dispatch), the notification is omitted and the conflicting plan name is set in the `conflict` attr in the state for usage within the plan component.

Patch-set: 5
Reviewer: Florian Fuchs <1006073@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=0