Update patch set 3
Patch Set 3:

(1 comment)

I got more clarification on how the toaster and viewed is supposed to work. The toaster should be dismissed after 8 secs unless the user is hovering then it is dismissed when the user stops hovering. 

Show the last notification received then previous notifications that were received (so newest first then remove and show older). There is still some debate on this...

Show the list of notifications and the number previously unviewed should be indicated by bg color. Once the notifications list is closed, mark all notifications in the list as viewed. I don't yet have the unviewed indication yet. That will be included in the next review.

Patch-set: 3