Update patch set 3
Patch Set 3:

(11 comments)

I would like to address the remaining comments in validations.less in the next review since there is some other rework there due to mockup changes.

I could kill this review and have these changes be in the next review, but that review would be that much bigger.  Sorry, I submitted this prior to the finalized mocks but I wanted to try to do smaller reviews.

Patch-set: 3