Adds API defaults and configuration.
Previously service IPs were hard-coded in src/js/APIEndpointUrls.js.
Users who used other IPs had to change them in this module and thus
change the source code of the application. This change intends to
ameliorate that.

The APIEndpointUrls module now looks up the IPs in a global object called
rdoDirectorUiConfig. If it cannot find an entry there, it will use the
current protocol and hostname plus the default ports and paths for
keystone, the tripleo api and the validations api. This default
behaviour is fine for all cases where the UI is installed directly
on the undercloud.

If the UI is installed on a different machine, users can create an
app.conf file (see app.conf.sample) in the root directory and add the
API urls there. Running gulp will read this file and generate a JS file
(./dist/js/rdo_director_ui_config.js) which is included in index.html
and sets the global rdoDirectorUiConfig object containing these URLs.

Change-Id: I4cab33114c0ddfee84cf8a5985fffc60e75c7e44