Reduxify Validations
Footer is now connected component providing also validations, Since the API
response with validation stages is nested of multiple entities, normalizr
and reselect are introduced. This combination enables to store those entities
(Stages, Validations and Results) to be stored separately in the state while
maintaining the relationships. Relesect is used to create selectors which
provide means to retrieve specific set/combination of data from the app state

Change-Id: I1af49d1fbdda93c0e881ae3645b567b9a132f84c